backend/requirements.txt
fastapi==0.110.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary
passlib==1.7.4
pyjwt==2.8.0
uvicorn==0.29.0
pydantic[email]
pytest
pytest-cov
pytest-asyncio
httpx